requirements.txt
boto==2.42.0
coreapi==1.32.0
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.8.10
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-filter==0.13.0
django-rest-auth==0.8.1
django-smart-selects==1.2.2
django-stdimage==2.3.3
django-storages-redux==1.3.2
django-tastypie==0.13.3
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
gunicorn==19.6.0
httpie==0.9.6
httpie-jwt-auth==0.2.1
itypes==1.1.0
Markdown==2.6.6
oauthlib==1.1.2
openapi-codec==1.0.0
Pillow==3.2.0
progressbar2==3.10.0
psycopg2==2.6.1
Pygments==2.1.3
PyJWT==1.4.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
python-utils==2.0.0
python3-openid==3.0.10
pytz==2016.6.1
requests==2.11.1
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
whitenoise==3.2
